Post fix
Post:
Now have avatars requests made for individual post for avatar population

Server:
now properly handles request for avatar. Was previously trying to use session, but that wasnt always set when a request for an avatar was made. Gets username from params of get request, and returns the avatar for that user.