OAuth added
User can now sign in through a google account

If the hashed google id already exists it will login to that account and redirect back to the home page. If not it will be asked to register and a new user will be added to the db with the returned hashed google id.

Users can logout, the session variables will be wipes and a logout request is sent to google to logout the user.